fix issue with cell_size and repeat_cell on tpms generate + add corresponding test
