+++ conda.recipe/meta.yaml
@@ -27,6 +27,7 @@ test:
   imports:
     - microgen
   commands:
+    - pip list
     - pip check
   requires:
     - pip
